make Collider mesh clone generic, add suzanne to test scene for collider mesh gen testing
