+++ Basis/Packages/com.basis.tests/Interactable/PickupInteractable.cs
@@ -26,19 +26,11 @@ public class PickupInteractable : InteractableObject
     private GameObject HighlightClone;
     private AsyncOperationHandle<Material> asyncOperationHighlightMat;
     private Material ColliderHighlightMat;
-    private bool previousKinematicValue = true;
+    private bool _previousKinematicValue = true;
     
     // constants
     const string k_LoadMaterialAddress = "Interactable/InteractHighlightMat.mat";
     const string k_CloneName = "HighlightClone";
-    private static Quaternion rotCapsuleX;
-    private static Quaternion rotCapsuleZ;
-
-    public PickupInteractable()
-    {
-        rotCapsuleX = Quaternion.Euler(new Vector3(0, 0, 90));
-        rotCapsuleZ = Quaternion.Euler(new Vector3(90, 0, 0));
-    }
 
     void Start()
     {
@@ -79,13 +71,10 @@ public class PickupInteractable : InteractableObject
 
         // NOTE: Collider mesh highlight position and size is only updated on Start(). 
         //      If you wish to have the highlight update at runtime do that elsewhere or make a different InteractableObject Script
-        HighlightClone = CloneColliderMesh(ColliderRef, gameObject.transform);
+        HighlightClone = ColliderClone.CloneColliderMesh(ColliderRef, gameObject.transform, k_CloneName);
 
         if (HighlightClone != null)
         {
-            HighlightClone.name = k_CloneName;
-            // generated meshes at this point should always have a meshrenderer
-
             if (HighlightClone.TryGetComponent(out MeshRenderer meshRenderer))
             {
                 meshRenderer.material = ColliderHighlightMat;
@@ -98,109 +87,7 @@ public class PickupInteractable : InteractableObject
 
     }
 
-    private GameObject CloneColliderMesh(Collider collider, Transform parent) {
-        GameObject primitive = null;
-        switch (collider.GeometryHolder.Type)
-        {
-            case GeometryType.Sphere:
-                var sphere = (SphereCollider)collider;
-                // TODO: use&cache sphere mesh generated (is lower poly)
-                primitive = GameObject.CreatePrimitive(PrimitiveType.Sphere);
-                if(primitive.TryGetComponent(out SphereCollider sCol))
-                {
-                    Destroy(sCol);
-                }
-                else
-                {
-                    BasisDebug.LogError("Primitive Sphere did not have a sphere collider?!");
-                }
-                primitive.transform.parent = parent;
-
-                primitive.transform.localPosition = sphere.center;
-                primitive.transform.localScale = Vector3.one * sphere.radius * 2;
-                break;
-            case GeometryType.Capsule:
-                var capsule = (CapsuleCollider)collider;
-                primitive = new GameObject(k_CloneName);
-                MeshFilter mFilter = primitive.AddComponent<MeshFilter>();
-                primitive.AddComponent<MeshRenderer>();
-                primitive.transform.parent = parent;
-
-                // generate mesh since we cant just scale the capsule primitve (sadly)
-                Mesh newMesh = CapsuleMeshGenerator.CreateCapsuleMesh(capsule.radius, capsule.height, 8);
-                mFilter.mesh = newMesh;
-
-                primitive.transform.localPosition = capsule.center;
-                
-                switch (capsule.direction)
-                {
-                    // X, Y (no change), Z
-                    case 0:
-                        primitive.transform.localRotation = rotCapsuleX;
-                        break;
-                    case 2:
-                        primitive.transform.localRotation = rotCapsuleZ;
-                        break;
-                    default:
-                        break;
-                }
-                primitive.transform.localScale = Vector3.one;
-
-                break;
-            case GeometryType.Box:
-                var box = (BoxCollider)collider;
-                primitive = GameObject.CreatePrimitive(PrimitiveType.Cube);
-                if(primitive.TryGetComponent(out BoxCollider boxCol))
-                {
-                    Destroy(boxCol);
-                }
-                else
-                {
-                    BasisDebug.LogError("Cube Primitve did not have a box collider?!");
-                }
-                primitive.transform.parent = parent;
-
-                primitive.transform.SetLocalPositionAndRotation(box.center, Quaternion.identity);
-                primitive.transform.localScale = box.size;
-                break;
-            case GeometryType.ConvexMesh:
-            case GeometryType.TriangleMesh:
-
-                if (!collider.TryGetComponent(out MeshFilter meshFilter))
-                {
-                    BasisDebug.LogWarning("Pickup Interactable with Mesh collider must have MeshFilter on same object to generate Highlight box");
-                    return null;
-                }
-                // mesh bounds, not collider bounds- as mesh bounds is in local space and not axis aligned
-                Bounds objectBounds = meshFilter.mesh.bounds;
-
-                primitive = GameObject.CreatePrimitive(PrimitiveType.Cube);
-                if(primitive.TryGetComponent(out BoxCollider _col))
-                {
-                    Destroy(_col);
-                }
-                else
-                {
-                    BasisDebug.LogError("Cube Primitve did not have a box collider?!");
-                }
-                primitive.transform.parent = parent;
-
-                primitive.transform.localPosition = objectBounds.center;
-                primitive.transform.localRotation = Quaternion.identity;
-                primitive.transform.localScale = objectBounds.size;
-                break;
-
-            // dont know how to handle remaning types 
-            case GeometryType.Terrain:
-            case GeometryType.Invalid:
-            default:
-                Debug.LogWarning("Interactable Object could not generate mesh collider highlight due to invalid collider type: " + collider.GeometryHolder.Type);
-                break;
-        }
-        
-        primitive.SetActive(false);
-        return primitive;
-    }
+    
 
 
     public void HighlightObject(bool highlight)
@@ -254,7 +141,7 @@ public class PickupInteractable : InteractableObject
 
             if(RigidRef != null && KinematicWhileInteracting)
             {
-                previousKinematicValue = RigidRef.isKinematic;
+                _previousKinematicValue = RigidRef.isKinematic;
                 RigidRef.isKinematic = true;
             }
 
@@ -278,7 +165,7 @@ public class PickupInteractable : InteractableObject
 
             if(KinematicWhileInteracting && RigidRef != null)
             {
-                RigidRef.isKinematic = previousKinematicValue;
+                RigidRef.isKinematic = _previousKinematicValue;
             }
